+++ lib/ecto/changeset.ex
@@ -257,9 +257,9 @@ defmodule Ecto.Changeset do
 
   ### Redacting fields in inspect
 
-  To hide a fields value from the inspect protocol of `Ecto.Changeset`, mark the field as `redacted: true`
-  in the schema, and it will display with the value `**redacted**`.
-
+  To hide a fields value from the inspect protocol of `Ecto.Changeset`, mark
+  the field as `redacted: true` in the schema, and it will display with the
+  value `**redacted**`.
   """
 
   require Ecto.Query